% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemiology-periods.R
\name{dissertation_covid_welle2_periods}
\alias{dissertation_covid_welle2_periods}
\title{Frozen dissertation principal COVID Welle2 intervals}
\usage{
dissertation_covid_welle2_periods()
}
\value{
A reviewed period resource.
}
\description{
Frozen dissertation principal COVID Welle2 intervals
}
