% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{dissertation_mean_age_spec}
\alias{dissertation_mean_age_spec}
\title{Dissertation mean-age specification}
\usage{
dissertation_mean_age_spec()
}
\value{
A validated indicator specification.
}
\description{
References the source-provided Regionaldatenbank mean age calculated from
all available individual ages. No mean-age calculation is performed.
}
