% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{dissertation_youth_dependency_ratio_spec}
\alias{dissertation_youth_dependency_ratio_spec}
\title{Dissertation youth dependency ratio specification}
\usage{
dissertation_youth_dependency_ratio_spec()
}
\value{
A validated indicator specification.
}
\description{
Preserves population aged 0--19 divided by population aged 20--64,
multiplied by 100. This function defines the specification but performs no
calculation.
}
