% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-context.R
\name{validate_context_base}
\alias{validate_context_base}
\title{Validate generic contextual base quantities}
\usage{
validate_context_base(x)
}
\arguments{
\item{x}{A data frame or tibble with \code{geo_id}, \code{geo_name}, \code{geo_level},
\code{geo_vintage}, \code{reference_date}, \code{variable}, \code{value}, \code{unit}, and \code{source}.
Required fields must be complete.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate generic contextual base quantities
}
