% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-typology-specs.R
\name{validate_dynamic_fitting_spec}
\alias{validate_dynamic_fitting_spec}
\title{Validate a dynamic typology fitting specification}
\usage{
validate_dynamic_fitting_spec(x)
}
\arguments{
\item{x}{A named dynamic fitting specification.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a dynamic typology fitting specification
}
